requirements.txt
Pillow==10.0.1
Pillow==10.1.0
streamlit==1.27.2
